cambios

+++ src/api/user.api.js
@@ -26,5 +26,5 @@ export const removeReport = (userId, reportId) => authApi.delete(`/${userId}/del
 export const getUserModules = (userId) => authApi.get(`/${userId}/get-modules/`);
 export const getUserReports = (userId) => authApi.get(`/${userId}/get-reports/`);
 export const getUserFavorites = (userId) => authApi.get(`/favorites/${userId}/`);
-export const addFavorite = () => authApi.post(`/favorites/`);
-export const removeFavorite = () => authApi.delete(`/favorites/`);
+export const addFavorite = (userId, reportId) => authApi.post(`/favorites/`, userId, reportId);
+export const removeFavorite = (userId, reportId) => authApi.delete(`/${userId}/favorites/${reportId}`);
